move various aws helper functions into dedicated modules
