+++ cloudigrade/util/tests/test_aws_arn.py
@@ -1,0 +1,95 @@
+"""Collection of tests for ``util.aws.helper`` module."""
+
+import faker
+from django.test import TestCase
+
+from util.aws.arn import AwsArn
+from util.exceptions import InvalidArn
+from util.tests import helper
+
+
+class UtilAwsArnTest(TestCase):
+    """AwnArn class test case."""
+
+    def test_parse_arn_with_region_and_account(self):
+        """Assert successful account ID parsing from a well-formed ARN."""
+        mock_account_id = helper.generate_dummy_aws_account_id()
+        mock_arn = helper.generate_dummy_arn(account_id=mock_account_id,
+                                             region='test-region-1')
+
+        arn_object = AwsArn(mock_arn)
+
+        partition = arn_object.partition
+        self.assertIsNotNone(partition)
+
+        service = arn_object.service
+        self.assertIsNotNone(service)
+
+        region = arn_object.region
+        self.assertIsNotNone(region)
+
+        account_id = arn_object.account_id
+        self.assertIsNotNone(account_id)
+
+        resource_type = arn_object.resource_type
+        self.assertIsNotNone(resource_type)
+
+        resource_separator = arn_object.resource_separator
+        self.assertIsNotNone(resource_separator)
+
+        resource = arn_object.resource
+        self.assertIsNotNone(resource)
+
+        reconstructed_arn = 'arn:' + \
+                            partition + ':' + \
+                            service + ':' + \
+                            region + ':' + \
+                            account_id + ':' + \
+                            resource_type + \
+                            resource_separator + \
+                            resource
+
+        self.assertEqual(mock_account_id, account_id)
+        self.assertEqual(mock_arn, reconstructed_arn)
+
+    def test_parse_arn_without_region_or_account(self):
+        """Assert successful ARN parsing without a region or an account id."""
+        mock_arn = helper.generate_dummy_arn()
+        arn_object = AwsArn(mock_arn)
+
+        region = arn_object.region
+        self.assertEqual(region, None)
+
+        account_id = arn_object.account_id
+        self.assertEqual(account_id, None)
+
+    def test_parse_arn_with_slash_separator(self):
+        """Assert successful ARN parsing with a slash separator."""
+        mock_arn = helper.generate_dummy_arn(resource_separator='/')
+        arn_object = AwsArn(mock_arn)
+
+        resource_type = arn_object.resource_type
+        self.assertIsNotNone(resource_type)
+
+        resource_separator = arn_object.resource_separator
+        self.assertEqual(resource_separator, '/')
+
+        resource = arn_object.resource
+        self.assertIsNotNone(resource)
+
+    def test_parse_arn_with_custom_resource_type(self):
+        """Assert valid ARN when resource type contains extra characters."""
+        mock_arn = 'arn:aws:fakeserv:test-reg-1:012345678901:test.res type:foo'
+        arn_object = AwsArn(mock_arn)
+
+        resource_type = arn_object.resource_type
+        self.assertIsNotNone(resource_type)
+
+        resource = arn_object.resource
+        self.assertIsNotNone(resource)
+
+    def test_error_from_invalid_arn(self):
+        """Assert error in account ID parsing from a badly-formed ARN."""
+        mock_arn = faker.Faker().text()
+        with self.assertRaises(InvalidArn):
+            AwsArn(mock_arn)
